Switch to env based configuration for the spider
This env file contains the configuration for the Node JS part of
the program. Using this we can define different settings, depending
of the environment the spider is run in. This is already used in the
two defined start commands "start" and "debug". Further, one can specify
settings for the database. An example document (using unix environment
variables) is available in the repository under spider.env.demo. To
use the configuration, one should rename the file to .env and modify
it to match the desired setting (see also the stack.yml file for
the DB settings).