Set up logger and cleared issue with cheerio and winston
Logger winston and cheerios $.each ran into an infinite loop problem,
which was detected by a control algorithm in util.js (part of the stack),
however, the exception always got consumed by winston, since it was the
default logger for the errourness case. This fact kept the infinite loop
alive. For now, we should not log within $.each. This function should
only be used for collecting the data needed, logging should take place
after this.

See [#1]