Adapted database scheme to store links and mime types
This way the database is easier extensible if we decide to handle
more datatypes in the future (as of now, we are only handling
html and json). The links are necessary if we decide to create
a darknet search engine (e.g. to employ pagerank algorithm). Also
it can be helpful for further research, if we have access to
the link structure, changing over time.