Switch to a more modular project setting
We split up the spider into three distinct, orthogonal
modules. This way we can ensure division of tasks and
keep the code readable. Further we introduced sequelize
instead of the db moduel, which makes the db handling
easier and most importantly, better readable.
To enforce good readability and later on easy documentation
generation, eslint now enforces jsdoc comments and validates
them as well. Travis will be set up to generate the docs
in every build in a next iteration of the travis settings.