Rewrite the networking module
Due to the new, modularized structure, I had to change to a http library,
that supports promises. This lead to axios, which seems to be a pretty
small (also only 2 dependencies), but powerful library. The network module
now has a dedicated get function (just in case we need a post function anytime
in the future), as well as a response handler, which can be used to serialize
the responses.
Overall the documentation of the code has been improved over the process.