Set up sequelize to store data and support for init data
Thie commit adds support for init data in the form of
CSV documents. The format of the CSV has to be one entry
per cell. Other formats may work, but are not guaranteed to.
These entries are inserted into the database with a timestamp
of 0, which ensures, they will be the first that get scraped
and therefor start our DB fillin up gradually.