Event based communication system between network and DB
We now use an event based system to communicate new data between the
database and the network module. This enables us to easily implement
pushback on lots (Constant) of connections to the network. After
receiving a new page, it will be scraped and the new data will be
inserted by the conductor. This would us also enable to use a
worker pool for parallel handling of the requests (e.g. download).