Changes in structure - now with only one event
This change contains a change in the communication between
Network module and Conductor module. This structure is clearer
and less error prone. Further, a depth field has been added
to the Path database, such that we can cut off more easily and
and get an idea of the structure. Still unsolved is however,
how to get the linking information in a clean way.