Switch to own network promise library
Since I ran into an issue with the network, I switched
back to http/https.get, in order to debug the issue.
Since I found it so easy, I stuck with this solution
and extended the project with another library function,
providing us with a promise based network interface.
The current solution is working and already completed
a minor local scrape run with one initial URL. However
the applied base64 filtering might be too harsh.
I found that quite a few pages encode their css in base64,
which will be filtered out with the current solution.
[#2]