Refactor and adaption to enhanced DB schema
In order to ensure not 100 simultaneous requests to one host,
the network module now owns a pool of entries to download
sorted by a random number. This should ensure a good distribution
accross the available hosts. However, we ensure a maximum of 6
simultaneous connections and queue the other requests if already
6 connections are available.
Futher the network module now exclusively emits events and is not
invoked over and over again by the conductor to prevent OOM exceptions
and too deep recursion.
For the conductor, this means that we need to adjust to a set of
small event handlers and initialization code.
[#7, #6]