Refactored the conductor to fit into the new structure
This contains an update on the conductor which does now only react
when needed and is not involved except for orchestrating the process.
The database helper functions saw a few small changes to be preciser
and also reflecting all changes done to the DB model. Continuous
Downloads not yet running through and subject to debug, rest of the
functions are running correctly. (Continuous Download = fetching new
data from db if available and do not get in the way of the downloader)
Idea: The downloader should never have to wait for the DB, since the
DB is much faster and we want to use as much of the time downloading
as possible.
[#6,#7,#8]