Minor bugfixes after system integration tests
Ensure that there are in any case enough urls in the pool (tests
with the edge case of last url downloaded - no left in the pool failed).
This is now ensured with an extra check after inserting new URIs into the
DB, if we can fullfill the pools needs afterwards.
Filtering of relative uris improved after mismatching full urls in test.

[#2]