Multiple Tor instances to reduce load on exit nodes
Since we ran into the issue with one exit node and 100 concurrent
requests, we decided to use multiple instances instead of one
single instance. This should distribute the load more evenly onto
multiple exit nodes. As of now, we spin up 100 nodes, but cannot
yet connect to all of them. This issue will be tackled in the next
commit.