Initial setup of tor-router completed
The tor-router is now set up at spider startup. For now
this also requires a restart of the tor container, however
most of the time the whole stack will be spin up or taken
down in a whole. This introduces now 1/5th instances of the
maximal concurrent requests to the network of tor instances.
However, this commit does not yet manage to actually download
over those instances.