Integrated tor-router and switched to multiprocessing
We encountered an issue in which case the scraper was too slow
to handle the incoming responses from the tor-router, which
resulted in a loss of a lot of potential useful data. In order
to circumvent this issues, we switched to a multi process
design. This allows us to catch up with the tor-router.
The tor-router itself needed some bugfixes, which are
already are published in the official github repository
of the scraper. However, since we may need further adaptions,
we keep our fork of it in order to have complete freedom
on the changes that we make.