Reconfigure build and run to work with all components
The run and build of the spider had to be adjusted, such that
all components can be run and even debugged in the production
environment. This is necessary to find issues in a longer run,
that were not foreseen. In some cases it is enough to abort the
execution and restart the spider in debug mode, since the not yet
downloaded entries are reloaded from the database. This is only
possible since we chosen to do the selection of new entries randomized
but still deterministic (the random value in the db does not change).
Further, previous changes broke the build and run system,
therefor a few adjustments were necessary, especially to ensure inter
container communication in both, the production and debug environment.