Only get entries from db that can be actually loaded
This commit fixes [#16].
It contains a small change in the indexing of the pending
request obects: The base url was replaced by the databases
baseUrlId, such that we can exclude corresponding entries
without another access to the database.
If we kept the baseUrl as before, we would either have to keep
a mapping from baseUrls to ids in memory or do a lookup on the
baseUrl index. The lookup would not be too bad, however, we can
do without it. The base url was never needed except for indexing.