A few optimizations on the database scheme and insertion
Link insertion is now solved by bulk inserting, which reduces the load
of the db by approximately 1/3. Further, the getter for the pool was
optimized such that it now only takes two requests to the db to
repopulate the pool. I found that the sorting information was missing
on the index for the repopulation of the code, which should result in
better performance.
This closes #17, since sequlize has not yet merged suport for bulk
upsert, which probably could be used for further optimizations when
inserting paths.