Updates packages and deactivates sequelize log
Note that the sequelize log only writes out the sql
queries that are sent to the db. This slows down the
performance, since we always write a lot of content out
to the terminal/log files, which are not relevant.
Therefor we introduce a dummy function which just ignores
the log output of sequelize, since there are not more
fine-grained options of logging in sequelize.