Support for multiple spiders added
This commit adds support for multiple spiders working on the same
dataset. This means, they do not interfere with each other and
one can specify, that a spider will be attached to an already
running session. That way, we do not set back the inProgress flags
and therefor do not double the work.
Close #19