Minor bugfixes in the handling of secure uris and constraints
For secure URIs (https) the secure flag was not set correctly,
it was set to "s". For the baseUrls we had an issue with
conflicting constraints: The baseUrl was set to unique on its
own and then we had a unique index on baseUrl and subdomain.
The later would have been correct, but both were active.
Corrected in this commit.