Serialize DB access
This was done in order to circumvent an issue in sequelize, that from
time to time (especially when vacuum processes are running) the connection
times out. In order to prevent that, this commit experiments with sequential
DB write. However, this is awfully slow, and I won't push it to the master,
since I think this is not a viable solution. I'll tinker some more with
the settings of Postgres and Sequelize and hope I can resolve the issue