Minor improvements and bugfixes after reviewing the logs
The logs show that in some cases the scraper starved, since
the error handler did not return the slot. Further, in some
cases the cheerio base url getter run into an issue in which
the scraper ran for a few minutes at 100% processor usage
and dropped quite a few responses. This can be circumvented
by just ignoring the base url tag. This should not hurt
since we know the host already at this point.