Introduces denormalized column for efficient prioritization
We introduced an incoming column of links, such that we scrape the most
important pages first. This column is important, since it is not
feasible to calculate the data dynamically (even after optimization of
the request, indexes and similar, it takes up to 1.5 minutes, if we
want to get 2000 entries). If we update only the paths we just found,
it is much cheaper (typically only a few dozen entries to update), and
since we can sort cheaply (index on the column) and limit fast, this
small amount of storage seems well spent space in contrast to the time
savings.
[#20]