Improve prioritizing algorithm - black hole defense
The higher complexity led to several bugs in the code, which took
some time to fix. We also got rid of the EventEmitter, in order
to prevent memory overruns and uncontrolled getters. We now use our
own callback queue, where we put the waiting clients. We'll try
to run this thing for some time on the server, to see how it performs.
I hope that with these changes we wont run again into performance
issue on the DB side. This should not happen, since we do not allow
more than the available connections to the database.
Close #20