Removed unnecessary condition on random in prioritization
This condition did not behave as expected and led to a condition
where for some baseUrls no paths were returned, even though they never
have been scraped before. This is now solved by just removing this
condition. This might lead to a condition, where we scrape several
paths from one host, but this is ok