Improved naming and prioritization (again)
This introduces a new setting in the .env file, which allows
the user to easily switch between the scraping modes (prioritization
of entries to download). We now also support a mode for inverse
prioritized scraping (the page with the least links is considered
the most important one, since it might bring much more additional
information). Naming and sorting in some queries is now expliciet,
which should make it clearer and the log volume is now also named
according to the naming scheme used for the other volumes
and containers.