First version of language detection without storage
This is a running example with the franc-min set, since
this seems to be preciser in our case, since we do not always
have large strings, it works better to get down to 82 languages.
Language classification works, however, the results are not yet
written to the database, in order to ensure no corruption of the
productive databases. I will thoroughly test the functionality
first before writing. Possibility of changes to the structure
after a review with @dionyziz.