Setting better default value for network failures
This commit addresses an issue in the scraper, where previously
always a general client error was assumed if the download failed.
Since this makes it hard to distinguish between failure report
from a server and actually failed (ev. offline) download, we now
set -1 as default value if the download runs into an exception (e.g.
Timeout).

Other changes included in this commit: changes in the database structure
of the preprocessing in order to allow for all possible relations.
Adaptions in the storage process not yet fully incorporated.