Adaptions to the scheme for data processing
This commit changes the datascheme to improve the ml performance
as well as it simplifies the handling of the data using sequelize.
Further it adapts the store logic used to persistently store the
entries in the database. However, this store logic is still buggy,
since it has some missing constraints.