Improved logging for textextraction
It seems that the textextraction for pdfs never works and for some
reason I do not see any logs in the logfiles and on the stdout about
it, therefor added a console.log to see if the logger is faulty or
it does not enter the exception handler.
Further added an apt update in the dockerfile, should help with an
installing issue of xpdf (first build it run through).