Adapted extractor to match new status code defaults
This includes that no -1 status coded contents are used for link
extraction (which would not make sense, but increase load on the db
as well as load on the machine). Further, removed very slow where
clauses, which in most cases do not make sense. It is faster to
shortly parse those empty files, than waiting for the db to sort
those out.