Removed timeout for waiting on DB connections (prep)
Since we have an upper bound on the number of connections to the db, we
also have an upper bound for the number of pending promises. Therefor
we should not run into memory leak issues, since we wait (semaphore-
like) for at least one promise to be resolved (reject terminates the
execution) before we aquire a db connection. This means that we might
wait a long time or starve if the db is offline. However, the db is
monitored and therefor detected to be offline likely before we see
that the preprocessor has crashed. Therefor we can safely remove
the timeout (also reducing the complexity (no null accesses)
of the whole structure).