Revert back to working sample of preprocessor&speed up build
The build was speed up by packing apt installs in a base image for the
preprocessor. This step took always quite some time on keftes, therefor
putting into an image that is less often built.

The change back was introduced, since the database was not able to
produce decent speeds with the better suited query, even after
fiddeling around with indexes, improved requests and reading tons of
documentation. Be warned: do not try to optimize this. Contents is
an issue for any database system, since it has large contents, which
makes it very expensive to do certain queries.