Added helpers, fixes several minor bugs and improves modularity
This commit contains quite a lot of changes. If on one hand fixes a few
bugs in the classifier and the datamodel, as well as improved
modularity in both, the models and the classifier, such that we have
no code duplication.
Second, improved usability with added configuration to .env as well
as command line options for the classifier. This allows us to switch
modes, add labelled data, train or apply models as well as define some
coarse constants for the training process, which makes it easy to
adapt the training process without having to restart the docker
container every time. However, applying it to a larger dataset
still has its issues, since the classifiert uses quite some memory,
which on smaller machines tends to lead to problems.
The memory usage stems from the large bag of word vectors, which
decreases performance considerably.
Todo: Check the classifier settings, in order to tweak performance
in both meanings: Training performance/memory usage as well as
precision of the resulting classifier.