Latest build does not install all necessary deps
On keftes we had the issue that not all necessary dependencies were
installed. However, in the dev environment it worked out perfectly.
The issue seemed to be, that node-svm was already installed on the
dev env whereas the deps for node-svm were not available on keftes.
The updated Dockerfile should now install all necessary deps.