Build folder misused by node-svm - .gitingore interfered
It did not build on keftes nor tik due to the fact that a directory was
missing. It was missing due to our gitignore, which factually is
correct, however the node-svm module does things slightly differently
and has a build/Release directory for some of its dependencies.
Solution: force add this to the repo