Improved performance of classifier enourmously
This should help us to get way better performance from the classifier,
since a) it takes less time to load data from disk and b)
the normalization of the feature vectors should help
to get better classification results.
Additionally added some convenience functionality to faster adjust
settings of the classifier and the preprocessing.