Correct access rights to volume
Ran into an issue on keftes, where the classifier was not able to write
the results into the files on a volume. This was caused by mounting the
volume as root, then changing to an unpriviledged user. Found a
workaround, that creates the mounting directory first, changes access
rights and then mount the volume on boot up.

Further the code was cleaned up considerably and error handling was
improved. No uncontrolled crashes should now happen and the correct
error messages should be displayed (in the order the appeared).