References now working, still unknown command
It appears that some styles and some versions of LaTex do not like
spaces in the citation command. It now compiles as bibtex and does
include the necessary references. However, one issue still remains
in the compilation with pdflatex, which I was not yet able to resolve.
This fixes the first part of issue #27