Minor revisions on the whitepaper
Explanation for the color change in the images describing legality:
It was brought to attention that especially american readers might
be confused of our red and blue color choice. To ensure no one
understands this for a subversive political message, we'll change
the color to something that should not ressemble any political
movement known to us.