Initialize structures required for multiple runs
This commit initializes the structure such that we can update the
sequelize models continuously. This allows us in the first place to
streamline the updates necessary to the db, required that we can
run several runs on the same database. That enables further that
we can compile all runs into one dataset instead of splitting it
up into multiple databases. The current commit is untested and
especially the migrations require some testing before we can
employ them.

For details on the changes required, please refer to issue #38.