Adjust to use sequelize migrations
This does not work without modification on existing database systems.
Once the patches are applied though, they wont be reapplied, hence
we can solve this issue by applying empty migrations (e.g. for the
foreign key adding migrations). This allows then to further use new
migrations to apply it to the db.