Added lazy robots check
Instead of eagerly loading and checking the robots.txt, we now only
load the file if needed. The provided function either gets the data
from the db if available, otherwise it starts a prioritized download
to get the file. Once the download or the db load has finished, the
file gets parsed and used to determine the validity of the given
dbResult object.